+++ include/Tickets_vect.h
@@ -4,7 +4,7 @@
 #include "Vectr.h"
 #include "Ticket.h"
 #include <iostream>
-#include <boost/make_shared.hpp>
+//#include <boost/make_shared.hpp>
 
 class Tickets_vect
 {
@@ -14,7 +14,7 @@ class Tickets_vect
         int get_size () const { return all.size(); }
 
         void add(std::shared_ptr<Ticket> one) { all.push_back(one); }
-        int find(int num_trans, int num_doc);   //поиск по номеру паспорта
+        int find(int num_trans, unsigned long num_doc);   //поиск по номеру паспорта
         const Ticket & get(int which) {return *(all[which]);}    //исключение генерируется автоматически
 
     protected:
